+++ lib/model/cross_file.dart
@@ -1,3 +1,5 @@
+import 'dart:typed_data';
+
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:localsend_app/model/file_type.dart';
 import 'package:wechat_assets_picker/wechat_assets_picker.dart';
@@ -12,6 +14,7 @@ class CrossFile with _$CrossFile {
     required String name,
     required FileType fileType,
     required int size,
+    required Uint8List? thumbnail,
     required AssetEntity? asset, // for thumbnails
     required String? path,
     required List<int>? bytes, // if type message, then UTF-8 encoded