+++ lib/widget/sliver/sliver_pinned_header.dart
@@ -1,0 +1,45 @@
+import 'package:flutter/material.dart';
+
+class SliverPinnedHeader extends StatelessWidget {
+  final double height;
+  final Widget child;
+
+  const SliverPinnedHeader({required this.height, required this.child});
+
+  @override
+  Widget build(BuildContext context) {
+    return SliverPersistentHeader(
+      pinned: true,
+      delegate: _SliverPersistentHeaderDelegate(
+        height: height,
+        child: child,
+      ),
+    );
+  }
+}
+
+class _SliverPersistentHeaderDelegate extends SliverPersistentHeaderDelegate {
+  final double height;
+  final Widget child;
+
+  _SliverPersistentHeaderDelegate({
+    required this.height,
+    required this.child,
+  });
+
+  @override
+  double get minExtent => height;
+
+  @override
+  double get maxExtent => height;
+
+  @override
+  Widget build(BuildContext context, double shrinkOffset, bool overlapsContent) {
+    return SizedBox.expand(child: child);
+  }
+
+  @override
+  bool shouldRebuild(_SliverPersistentHeaderDelegate oldDelegate) {
+    return false;
+  }
+}